Migrated ScopeFileSystemsTests to integration tests
